+++ samples/kotlin-maven-simple/src/main/kotlin/org/refactorkit/samples/Greeting.kt
@@ -4,3 +4,10 @@ class Greeting
 interface GreetingPort
 enum class GreetingMode { HELLO }
 annotation class GreetingMarker
+object GreetingRegistry
+data object GreetingDataRegistry
+class GreetingOwner {
+    companion object
+    object NestedRegistry
+    val anonymous = object {}
+}